add Rate

+++ app/src/main/java/com/example/napat/movieapp/model/Rate.kt
@@ -2,5 +2,5 @@ package com.example.napat.movieapp.model
 
 data class Rate(
         var id : Int,
-        var ratingPoint : ArrayList<Int>
+        var ratingPoint : ArrayList<Float>
 )
